Move functions from `type-spec` to `integration`
